update session funcitonality
